refactor(FirebasePlugin.java): Clear launch notification once requested
Ensure parity with iOS implementation.

Also fixes an issue where, when launchNotification handling happens on webview load and the webview is reloaded, the app would behave as if the user had tapped on the notification again instead of just starting normally.

Ideally, this behavior should be moved to handling of webview reload